handle cpal errors before starting playback (#16)
* allow passing gp5 file as cli argument

this is a quick fix that was added because i don't use a desktop
environment on linux, and wasn't able get the file picker to appear,
despite having installed several xdg-desktop-portal-* packages

* check cpal device/config before calling new_output_stream

cpal's device.default_output_config() may panic in edge cases, e.g. when
using USB speakers. if this happens, playback should not be started.

* fix fn arg formatting

* minor bugfixes